refactor: place nav drawer closer to notes ui
