+++ client/src/layout/NotesLayout/ui/NavigationDrawer/index.ts
