+++ client/src/layout/NotesLayout/ui/NavigationDrawer/ui/BackButton.tsx
