+++ client/src/layout/NotesLayout/ui/NavigationDrawer/ui/LabelsList.tsx
@@ -1,7 +1,7 @@
 import { routes } from '@shared/lib/const';
 import NavItem from '@shared/ui/NavItem';
 
-import { mockLabels } from '../../../../dev-data';
+import { mockLabels } from '../../../../../../dev-data';
 
 const LabelsList = () => {
   return (
